+++ src/Lib/PdfWriter/OrderListByProductPdfWriter.php
@@ -14,73 +14,7 @@
  */
 namespace App\Lib\PdfWriter;
 
-use App\Lib\Pdf\ListTcpdf;
-use Cake\Core\Configure;
-use Cake\Datasource\FactoryLocator;
-
-class OrderListByProductPdfWriter extends PdfWriter
+class OrderListByProductPdfWriter extends OrderListPdfWriter
 {
 
-    public $Manufacturer;
-
-    public function __construct()
-    {
-        $this->plugin = 'Admin';
-        $this->setPdfLibrary(new ListTcpdf());
-        $this->Manufacturer = FactoryLocator::get('Table')->get('Manufacturers');
-    }
-
-    public function prepareAndSetData($manufacturerId, $pickupDay, $validOrderStates, $orderDetailIds)
-    {
-
-        $manufacturer = $this->Manufacturer->find('all', [
-            'conditions' => [
-                'Manufacturers.id_manufacturer' => $manufacturerId
-            ],
-            'contain' => [
-                'AddressManufacturers'
-            ],
-        ])->first();
-
-        $productResults = $this->Manufacturer->getDataForInvoiceOrOrderList(
-            $manufacturerId,
-            'product',
-            $pickupDay,
-            null,
-            $validOrderStates,
-            Configure::read('appDb.FCS_INCLUDE_STOCK_PRODUCTS_IN_INVOICES'),
-            $orderDetailIds,
-        );
-
-        // calculate sum of price
-        $sumPriceIncl = 0;
-        $sumPriceExcl = 0;
-        $sumTax = 0;
-        $sumAmount = 0;
-        $sumTimebasedCurrencyPriceIncl = 0;
-        foreach ($productResults as $result) {
-            $sumPriceIncl += $result['OrderDetailPriceIncl'];
-            $sumPriceExcl += $result['OrderDetailPriceExcl'];
-            $sumTax += $result['OrderDetailTaxAmount'];
-            $sumAmount += $result['OrderDetailAmount'];
-            if (isset($result['OrderDetailTimebasedCurrencyPriceInclAmount'])) {
-                $sumTimebasedCurrencyPriceIncl += $result['OrderDetailTimebasedCurrencyPriceInclAmount'];
-            }
-        }
-
-        $this->setData([
-            'productResults' => $productResults,
-            'manufacturer' => $manufacturer,
-            'currentDateForOrderLists' => Configure::read('app.timeHelper')->getCurrentDateTimeForFilename(),
-            'sumPriceIncl' => $sumPriceIncl,
-            'sumPriceExcl' => $sumPriceExcl,
-            'sumTax' => $sumTax,
-            'sumAmount' => $sumAmount,
-            'sumTimebasedCurrencyPriceIncl' => $sumTimebasedCurrencyPriceIncl,
-            'variableMemberFee' => $this->Manufacturer->getOptionVariableMemberFee($manufacturer->variable_member_fee),
-        ]);
-
-    }
-
 }
-
